cool#9992 lok doc sign: fix signature status after load
Load a document, see the 'signaturestatus: 4' callback
(SignatureState::NOTVALIDATED, "signature is OK, but certificate could
not be validated"), wait till the view is initialized, notice the
strange later arriving 'statechanged: .uno:Signature=3' callback, which
hints that the document was modified, but we're right after load, so
that should not happen.

I already tried to prevent doc modified status in commit
654f972a97c374fa90eb3984a44c7f54ddba9c61 (sw lok: make sure author name
change doesn't mark the doc as modified, 2023-01-16), but this is not
enough for the doc sign case, where SfxObjectShell::SetModified() ends
up refreshing the doc title, which calls
SfxObjectShell::GetDocumentSignatureState() and then caches the wrong
SignatureState::INVALID state.

Fix the problem by adding a flag to not mark the document as modified
during "init view" in the first place: that keeps the document unchanged
after updating fields, but doesn't break the doc sign status.

Now LOK clients get a first 'signaturestatus: 4' on load, and a
'statechanged: .uno:Signature=4' on initializing the view, which is a
bit redundant, but at least is consistent.

(cherry picked from commit a95b90bc0cf0f8a58fc20684acc23bb89b827cd1)

Conflicts:
	sw/inc/viewsh.hxx
	sw/qa/extras/tiledrendering/data/signed-doc.odt
	sw/qa/extras/tiledrendering2/tiledrendering2.cxx

Change-Id: I42d50525ba96900eeeb927a50f1484c882347b13
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173584
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>