tdf#163152 don't convert image's sRGB colorspace
With Skia/Raster or Skia disabled, converting the image's
colorspace to match the window's colorspace causes more than
an expected amount of color saturation so let the window's
underlying CGContext handle any necessary colorspace conversion
in CGContextDrawImage().

With Skia/Metal, this bug is caused by the CAMetalLayer being
set to the same colorspace as its matching window. So set the
CAMetalLayer's colorspace to sRGB so that, like with Skia/Raster
and Skia disabled, any colorspace conversion is handled when the
CAMetalLayer is drawn to the window.

Change-Id: Ifa2abe46d34bfcf5acd478fffd346603f869157b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173962
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
(cherry picked from commit e4ab68142c7bc4e04ffe429567dda974b86985a7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174005
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>