Revert "Return earlier if there are no proposals from Duden checker."
This reverts commit bf980c7b7c8197724cbebd1ed16dfe4401b97392.
The reason is: the change caused empty entries in the ProofreadingResult's
aErrors; those didn't even have nErrorType set, and it was 0. Later, in
SwXTextMarkup::commitMultiTextMarkup, the value of the type was checked,
and when it was neither SENTENCE nor PROOFREADING, the application of the
result to the paragraph was aborted. Thus, any paragraph with a spelling
error without a suggestion wasn't marked as a whole.

A better fix is in works.

Change-Id: Ic5a2c75bb3896f34e60589d5fcbc8e9048dcda9b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175021
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>